src/requirements.txt
alabaster==1.0.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
binaryornot==0.4.4
boto3==1.35.39
botocore==1.35.39
bs4==0.0.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==43.0.1
cssbeautifier==1.15.4
cssselect==1.2.0
defusedxml==0.7.1
dj-database-url==3.0.1
Django==4.2
django-allauth==65.0.2
django-browser-reload==1.16.0
django-celery-beat==2.7.0
django-cleanup==9.0.0
django-cotton==2.1.3
django-prometheus==2.3.1
django-storages==1.14.4
django-tailwind==3.8.0
django-timezone-field==7.0
django-yamlfield==1.2.2
djangorestframework==3.15.2
djlint==1.36.4
dnspython==2.7.0
docutils==0.21.2
EditorConfig==0.17.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
incremental==24.7.2
influxdb-client==1.47.0
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jinja2-time==0.2.0
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.0
kombu==5.4.2
livereload==2.7.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
myst-parser==4.0.0
oauthlib==3.2.2
orjson==3.10.7
packaging==24.1
parsel==1.9.1
pathspec==0.12.1
pillow==10.4.0
pip-review==1.3.0
poyo==0.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
Protego==0.3.1
psycopg==3.2.3
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
reactivex==4.0.4
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.2
s3transfer==0.10.3
Scrapy==2.12.0
sentry-sdk==2.16.0
service-identity==24.2.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-autobuild==2024.10.3
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==1.0.0
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
starlette==0.39.2
text-unidecode==1.3
tldextract==5.1.3
tomli==2.0.2
tornado==6.4.1
tqdm==4.67.1
Twisted==24.10.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultimate-sitemap-parser==0.5
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
vine==5.1.0
w3lib==2.2.1
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
whitenoise==6.7.0
zipp==3.20.2
zope.interface==7.1.1
